+++ frontend/src/mocks/handlers.ts
@@ -1,35 +1,3 @@
-// src/mocks/handlers.js
-import { writingURL } from 'constants/apis/url';
-import { rest } from 'msw';
-import type { GetWritingResponse } from 'types/apis/writings';
-import { writingContentMock } from './writingContentMock';
+import { writingHandlers } from './handlers/writing';
 
-export const handlers = [
-  // 글 생성(글 업로드): POST
-  rest.post('/writings/file', async (_, res, ctx) => {
-    return res(ctx.status(201), ctx.set('Location', `/writings/1`));
-  }),
-
-  // 글 블로그로 발행: POST
-  rest.post('/writings/:writingId/publish', async (req, res, ctx) => {
-    const blog = ['MEDIUM', 'TISTORY'];
-    const { writingId } = req.params;
-    const { publishTo } = await req.json();
-
-    if (!blog.includes(publishTo) || typeof writingId !== 'number') return res(ctx.status(404));
-
-    return res(ctx.status(200));
-  rest.get(`${writingURL}/:writingId`, (req, res, ctx) => {
-    const writingId = Number(req.params.writingId);
-
-    return res(
-      ctx.delay(300),
-      ctx.status(200),
-      ctx.json<GetWritingResponse>({
-        id: writingId,
-        title: '테스트 글 제목',
-        content: writingContentMock,
-      }),
-    );
-  }),
-];
+export const handlers = [...writingHandlers];
